requirements.txt
python-dotenv
PyYAML
ipython
tkvideoplayer

# Requerimientos para inferencia YOLOv5 y YOLOv7

# Base ----------------------------------------
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.1
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
torch>=1.7.0,!=1.12.0
torchvision>=0.8.1,!=0.13.0
tqdm>=4.41.0
protobuf<4.21.3

# Logging -------------------------------------
tensorboard>=2.4.1
# wandb

# Export / Inference --------------------------------------
coremltools>=6.0  # CoreML export
onnx>=1.9.0  # ONNX export
onnx-simplifier>=0.4.1  # ONNX simplifier
nvidia-pyindex  # TensorRT export

# Plotting ------------------------------------
pandas>=1.1.4
seaborn>=0.11.0

# Extras
psutil

opencv-contrib-python